apps/api/pyproject.toml
[project]
name = "forensic-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.135.1",
    "google-genai>=1.67.0",
    "httpx>=0.28.1",
    "imagehash>=4.3.2",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.41.0",
]
